Network_Recon_2.0/requirements.txt
python-nmap
psutil
requests
dnspython
python-whois
python3-nmap
python-nmap
